Add GitHub link with version info to login page (#481)
* Initial plan

* Add GitHub link with version info to login page

Agent-Logs-Url: https://github.com/Doezer/Questarr/sessions/5783c409-fcab-437f-8b13-a3e9cd7baaad

* Update client/src/pages/auth/login.tsx


* Update client/src/pages/auth/login.tsx


* Update client/src/pages/auth/login.tsx


* Extract GitHubVersionLink shared component to avoid duplication

Agent-Logs-Url: https://github.com/Doezer/Questarr/sessions/83fa7b03-0864-43cc-a758-f527eb9f4874

* Fix sonar smell

---------
